% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{gather_t_profile}
\alias{gather_t_profile}
\title{Make a tidy data frame from time profile output e.g. for use with ggplot}
\usage{
gather_t_profile(x, ...)
}
\arguments{
\item{x}{A time profile as generated by \code{\link{t_profile}}.}

\item{...}{Additional argument passed to methods and eventually to
\code{\link[tidyr]{gather}}.}
}
\description{
Make a tidy data frame from time profile output e.g. for use with ggplot
}
\details{
The idea is that you can take results generated by
  \code{\link{t_profile}} (usually as a time series object) and turn them
  into a \bold{tidy} data.frame suitable for \code{tidyverse} style analysis.

  If you have a \code{list} of t_profile results generated for multiple
  separate experiments, you can gather those together as well to make a
  single grand summary data.frame.
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link{t_profile}}.
}
