% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperstack.R
\name{make_hyperstack}
\alias{make_hyperstack}
\alias{make_hyperstack_multifile}
\title{Read one or more TIFFs into 4D hyperstack}
\usage{
make_hyperstack(x, zslices, timepoints, time.axis = 4)

make_hyperstack_multifile(x, file.axis = 4)
}
\arguments{
\item{x}{Path to (one or more) TIFFs}

\item{zslices}{Number of zlices in each TIFF}

\item{timepoints}{Number of timepoints in each TIFF}

\item{time.axis}{Which axis will end up as the time axis in the output}

\item{file.axis}{Which axis will end up as the file axis in the output when
reading multiple files.}
}
\value{
a 4D array typically ordered x,y,z,t or x,y,t,file
}
\description{
\code{make_hyperstack} reads a single TIFF which may have 3-4
  dimensions.

\code{make_hyperstack_multifile} reads multiple TIFFs each of
  which will have 2-3 dimensions (usually x,y,t) and combines into a single
  hyperstack x,y,t,file
}
