% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_roi_from_mean}
\alias{make_roi_from_mean}
\title{Construct a mean image, blur and threshold to get}
\usage{
make_roi_from_mean(x, blur = 5, thr = "95\%", ...)
}
\arguments{
\item{x}{A hyperstack or path to one or more TIFF files (see
\code{\link{make_hyperstack}})}

\item{blur}{The gaussian sigma to use}

\item{thr}{A thresholding specification (passed to
\code{imager::\link[imager]{threshold}})}

\item{...}{Additional arguments passed to \code{\link{make_hyperstack}}}
}
\value{
A 2D mask in the form of an imager::cimg object
}
\description{
Construct a mean image, blur and threshold to get
}
\details{
This is currently targeted at single plane imaging data.
}
\seealso{
\code{\link[imager]{cimg}}, \code{\link{t_profile}},
  \code{\link{make_hyperstack}}
}
