% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read.any.tiff}
\alias{read.any.tiff}
\title{Convenience function to read any TIFF image with attached information}
\usage{
read.any.tiff(f, ...)
}
\arguments{
\item{f}{Path to TIFF}

\item{...}{Additional arguments passed to \code{\link{read.scanimage}} or
\code{\link[tiff]{readTIFF}}.}
}
\value{
An array of the dimensions height x width x channels. If there is
  only one channel the result is a matrix. The values are integers.
}
\description{
Convenience function to read any TIFF image with attached information
}
