% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read.scanimage}
\alias{read.scanimage}
\title{Read scanimage tiffs into an R object}
\usage{
read.scanimage(source, slices = Inf, channels = Inf, info = T,
  as.is = TRUE, ...)
}
\arguments{
\item{source}{Either name of the file to read from or a raw vector
representing the TIFF file content.}

\item{slices}{Which slices to return (default of Inf reads all available
slices)}

\item{channels}{which channels to read. FIXME only the default
(\code{channels=Inf}) implying read all channels is presently implemented.}

\item{info}{if set to \code{TRUE} then the resulting image(s) will also
contain information from TIFF tags as attributes}

\item{as.is}{Return original pixel values without rescaling where possible.
This is usually what you want for 8 and 16 bit scanimage TIFFs (default
\code{TRUE}).}

\item{...}{additional arguments for \code{\link[tiff]{readTIFF}}.}
}
\value{
An array of the dimensions height x width x channels. If there is
  only one channel the result is a matrix. The values are integers.
}
\description{
Read scanimage tiffs into an R object
}
\examples{
t=read.scanimage(system.file('extdata/Blank-IPA_1s_16r_032.tif', package='scanimage'))
}
