% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{scanimageinfo}
\alias{scanimageinfo}
\title{Extract key image information from a scanimage TIFF}
\usage{
scanimageinfo(x)
}
\arguments{
\item{x}{One or more paths to TIFFs or an R object containing a TIFF image}
}
\value{
A data.frame with one row for every TIFF image in \code{x} and
  columns

  \itemize{

  \item h number of rows (aka scanimage lines)

  \item w number of columns (aka scanimage pixels per line)

  \item nchan number of image channels

  \item nframes number of image frames (i.e. timepoints*Z slices)

  }
}
\description{
Extract key image information from a scanimage TIFF
}
\examples{
img=system.file('extdata/Blank-IPA_1s_16r_032.tif', package='scanimage')
scanimageinfo(img)
}
\seealso{
\code{\link{parse_description}}
}
