% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_description}
\alias{parse_description}
\title{parse description of ScanImage TIFF file, converting it to key-value list}
\usage{
parse_description(x, raw = FALSE)
}
\arguments{
\item{x}{Path to a TIFF file or a raw data block}

\item{raw}{Whether to return the raw description field as a single string or
when \code{FALSE} (the default) to return it is as a list containing parsed
R data types.}
}
\value{
a named \code{list} or, when \code{raw=TRUE}, a character vector of
  length 1.
}
\description{
parse description of ScanImage TIFF file, converting it to key-value list
}
\details{
ScanImage TIFF flies contain a single description field, which is a
  CR delimited values of the form \code{key=value}.
}
\examples{
desc=parse_description(system.file(
  'extdata/Blank-IPA_1s_16r_032.tif',package='scanimage'))
}

