% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanimage-package.r
\docType{package}
\name{scanimage-package}
\alias{scanimage}
\alias{scanimage-package}
\title{scanimage: A package to read 2 photon image data acquired by ScanImage}
\description{
scanimage: A package to read 2 photon image data acquired by ScanImage
}

