% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{t_profile}
\alias{t_profile}
\title{Construct a temporal profile of voxels within an ROI}
\usage{
t_profile(x, mask)
}
\arguments{
\item{x}{A hyperstack with imaging data}

\item{mask}{An ROI mask e.g. from \code{\link{make_roi_from_mean}}}
}
\value{
A matrix with the third and fourth dimensions of \code{x}
}
\description{
Construct a temporal profile of voxels within an ROI
}
\seealso{
\code{\link{make_roi_from_mean}}
}
